pyproject.toml
[project]
name = "spat-slozno"
version = "0.0.1a"
description = "Приложение для получения статистики о сне."
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "flask>=3.1.0",
    "flask-cors>=5.0.1",
    "psycopg2-binary>=2.9.10",
    "python-dotenv>=1.0.1",
]

[tool.pyright]
venvPath = "."
venv = ".venv"
